+++ libft/ft_printf/redirect_print.c
@@ -51,13 +51,13 @@ size_t	print_str_fs(t_fs *fs, va_list ap)
 			i += print_num(fs, ap);
 		else if (fs->ch == 'c' || fs->ch == 'C')
 			i += print_char(fs, ap);
-		i += padding_after(fs, i);
+		i += padding_after(fs, (int)i);
 	}
 	else
 	{
 		i += padding(fs, 1);
 		i += ft_putchar_pf(fs->ch);
-		i += padding_after(fs, i);
+		i += padding_after(fs, (int)i);
 	}
 	return (i);
 }
